+++ onecool_os/assets/funds/__init__.py
@@ -1,0 +1,11 @@
+"""Funds asset module."""
+
+from onecool_os.assets.funds.loader import FundImportResult, FundLoader
+from onecool_os.assets.funds.models import FundAsset, FundPosition
+
+__all__ = [
+    "FundAsset",
+    "FundImportResult",
+    "FundLoader",
+    "FundPosition",
+]